Add SIGNEXTEND for generic target (closes #24)

+++ src/main.rs
@@ -383,7 +383,7 @@ unsafe fn run_evm(bytecode: &[u8], rom: &VmRom, schedule: &Schedule, gas_limit: 
                 let value = *((stack.sp.offset(-1) as *const u8).offset(offset));
                 let a = stack.pop();
                 let b = stack.pop();
-                let result = signextend_u256(a, b, value as i64);
+                let result = signextend_u256(a, b, value);
                 stack.push(result);
                 //
                 pc += 1;
